Added 1 test to time_test.cpp named "convertDuration4Fine2WithBigNumbers" Added 1 test to duration_test.cpp named "convertBigNumbersMicrosecondsToDuration" Changed the data type of the transferred parameter of the Microseconds-constructor, Milliseconds-constructor etc. in the duration.h (int32_t to int64_t)

+++ modules/time/src/cobc/time/duration.h
@@ -282,9 +282,9 @@ class Seconds : public Duration
 {
 public:
     explicit inline
-    Seconds(int32_t value) :
-        Duration(static_cast<int64_t>(value) * numberOfMillisecondsPerSecond *
-                                               numberOfMicrosecondsPerMillisecond)
+    Seconds(int64_t value) :
+        Duration(value * numberOfMillisecondsPerSecond *
+                         numberOfMicrosecondsPerMillisecond)
     {
     }
 };
@@ -298,8 +298,8 @@ class Milliseconds : public Duration
 {
 public:
     explicit inline
-    Milliseconds(int32_t value) :
-        Duration(static_cast<int64_t>(value) * numberOfMicrosecondsPerMillisecond)
+    Milliseconds(int64_t value) :
+        Duration(value * numberOfMicrosecondsPerMillisecond)
     {
     }
 };
@@ -313,8 +313,8 @@ class Microseconds : public Duration
 {
 public:
     explicit inline
-    Microseconds(int32_t value) :
-        Duration(static_cast<int64_t>(value))
+    Microseconds(int64_t value) :
+        Duration(value)
     {
     }
 };
